Move source files into src folder
