+++ __tests__/lruCache.js
@@ -1,4 +1,4 @@
-import { LruCacheObject } from '../index';
+import { LruCacheObject } from '../src/index';
 
 function fillCache(cache, entries = []) {
   entries.map(entry => cache.set(entry, entry));
